Extract reviewChanges types
- Add shared command API types under src/types
- Keep reviewChangesCommand re-exports for compatibility
- Point README consumers at the shared types file
